Fixed Interrupt and debunced the button
Changed the interrupt to trigger on the falling edge of the trigger
rather than rising. This fixed an issues with the interrupt being
triggered when twice in a row. I also added a long deb ounce of 200ms
to the button press. This also helped in fixing the switch bouncing and
trigger the interrupt when it shouldn't have.